:twisted_rightwards_arrows: merge models and a bit of clean up
