+++ src/models/Article.cpp
@@ -8,11 +8,11 @@ using namespace std;
 Article::Article() : Review() {
     this->name = "Unknown";
 }
-Article::Article(const string& title, const string& author, bool borrowed,
+Article::Article(const string& title, const string& author,
                  int publicationYear, int numberPages, const string& genre,
                  const string& summary, const string& editor, int numberArticles,
                  const string& name)
-    : Review(title, author, borrowed, publicationYear, numberPages, genre, summary, editor, numberArticles) {
+    : Review(title, author, publicationYear, numberPages, genre, summary, editor, numberArticles) {
     this->name = name;
 }
