+++ src/models/DVD.cpp
@@ -5,7 +5,7 @@ using namespace std;
 // Constructors
 DVD::DVD() : CD() {
 }
-DVD::DVD(const string& title, const string& author, bool borrowed,
+DVD::DVD(const string& title, const string& author,
          int secondesDuration, int numberTracks, const string& productionCompany)
-    : CD(title, author, borrowed, secondesDuration, numberTracks, productionCompany) {
+    : CD(title, author, secondesDuration, numberTracks, productionCompany) {
 }
