+++ src/models/Review.h
@@ -11,7 +11,7 @@ class Review : public Book {
         using Book::Book; // héritage de Book
         // Constructors
         Review();
-        Review(const std::string& title, const std::string& author, bool borrowed,
+        Review(const std::string& title, const std::string& author,
                int publicationYear, int numberPages, const std::string& genre,
                const std::string& summary, const std::string& editor, int numberArticles);
         // Getters
